Fix search worker build configuration and some bugs.
